+++ packages/common/utils/pattern-match.js
@@ -1,7 +1,7 @@
 const micromatch = require("micromatch")
 
 module.exports = (str, pattern, options = {}) =>
-  micromatch.isMatch(str, pattern, {
+  micromatch.match(str, pattern, {
     regex: true,
     ...options,
-  })
+  }).length > 0