+++ plugins/contrib/patches/05.1-needs-using-initcontainers.js
@@ -1,5 +1,4 @@
-const kontinuousNeedsImage =
-  "ghcr.io/socialgouv/kontinuous/wait-needs:v1.109.13"
+const kontinuousNeedsImage = "ghcr.io/socialgouv/kontinuous/wait-needs:v1.111.6"
 // const kontinuousNeedsImage = "harbor.fabrique.social.gouv.fr/sre/kontinuous/wait-needs:v1"
 
 const getDeps = require("../lib/get-needs-deps")