+++ CHANGELOG.md
@@ -2,6 +2,144 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.111.6](https://github.com/socialgouv/kontinuous/compare/v1.111.5...v1.111.6) (2023-01-10)
+
+
+### Bug Fixes
+
+* dependabot secu alerts ([0f328e2](https://github.com/socialgouv/kontinuous/commit/0f328e265bac9cdd19dfd8df096e143290e6b5eb))
+* namespace name ([80a8358](https://github.com/socialgouv/kontinuous/commit/80a8358baedf74f525e86684e37b352d24637ea6))
+
+### [1.111.5](https://github.com/socialgouv/kontinuous/compare/v1.111.4...v1.111.5) (2023-01-09)
+
+
+### Bug Fixes
+
+* build tests snapshots ([b648a72](https://github.com/socialgouv/kontinuous/commit/b648a72de695421183bd014db0497f55f16b81ef))
+* up rollout-status ([3d075e9](https://github.com/socialgouv/kontinuous/commit/3d075e9e0bbf0fdc9bfd76e98c563d99b92ca50f))
+
+### 1.111.4 (2023-01-09)
+
+
+### Bug Fixes
+
+* more retries for jobs ([#121](https://github.com/socialgouv/kontinuous/issues/121)) ([0ecec04](https://github.com/socialgouv/kontinuous/commit/0ecec04a530b6f957c562df5173f829f84d191e2))
+
+### [1.111.3](https://github.com/socialgouv/kontinuous/compare/v1.111.2...v1.111.3) (2023-01-09)
+
+
+### Bug Fixes
+
+* improve retry rollout-status debugging ([9780619](https://github.com/socialgouv/kontinuous/commit/97806191c7e36b62e3fa77789e946907b6c3b5e9))
+* log retry count ([219f9ee](https://github.com/socialgouv/kontinuous/commit/219f9eedf62ad0fcac5a928577ae5118663ffba6))
+
+### [1.111.2](https://github.com/socialgouv/kontinuous/compare/v1.111.1...v1.111.2) (2023-01-09)
+
+
+### Bug Fixes
+
+* improve retries ([01076f5](https://github.com/socialgouv/kontinuous/commit/01076f520a82b90ba396e65fdf4209093ad29590))
+
+### [1.111.1](https://github.com/socialgouv/kontinuous/compare/v1.111.0...v1.111.1) (2023-01-09)
+
+
+### Bug Fixes
+
+* async rollout-status throw-retriable ([74acac9](https://github.com/socialgouv/kontinuous/commit/74acac9fe0de86c32b13be6519343b339eee29a5))
+
+## 1.111.0 (2023-01-07)
+
+
+### Features
+
+* config plugin ifenv ([#120](https://github.com/socialgouv/kontinuous/issues/120)) ([2456d39](https://github.com/socialgouv/kontinuous/commit/2456d39ea5acfd1fea3750aa0b7dc27092d0eead))
+
+### [1.110.5](https://github.com/socialgouv/kontinuous/compare/v1.110.4...v1.110.5) (2023-01-06)
+
+
+### Bug Fixes
+
+* up rollout-status ([7cf9d7c](https://github.com/socialgouv/kontinuous/commit/7cf9d7c8561ce5fbbd9930ba5168257ad78bc92b))
+
+### [1.110.4](https://github.com/socialgouv/kontinuous/compare/v1.110.3...v1.110.4) (2023-01-06)
+
+
+### Bug Fixes
+
+* rollout-status retry status ([9c82ed5](https://github.com/socialgouv/kontinuous/commit/9c82ed572e57c497293112a05360d6189b2624b6))
+
+### [1.110.3](https://github.com/socialgouv/kontinuous/compare/v1.110.2...v1.110.3) (2023-01-06)
+
+
+### Bug Fixes
+
+* improve immutable detection ([3c4a083](https://github.com/socialgouv/kontinuous/commit/3c4a08375ba7e6dd114b6f7321c882fdd88ee597))
+
+### [1.110.2](https://github.com/socialgouv/kontinuous/compare/v1.110.1...v1.110.2) (2023-01-06)
+
+
+### Bug Fixes
+
+* retry ErrImagePull ([10e2744](https://github.com/socialgouv/kontinuous/commit/10e2744cb36cfe8552bd6862f9c8995caf9839f2))
+
+### [1.110.1](https://github.com/socialgouv/kontinuous/compare/v1.110.0...v1.110.1) (2023-01-06)
+
+
+### Bug Fixes
+
+* typo status.error.type ([4d610ec](https://github.com/socialgouv/kontinuous/commit/4d610ec6fc9e44953c88a50faa0940fef27d6744))
+
+## [1.110.0](https://github.com/socialgouv/kontinuous/compare/v1.109.17...v1.110.0) (2023-01-05)
+
+
+### Features
+
+* refactor rollout-status callers ([4086232](https://github.com/socialgouv/kontinuous/commit/40862329925a8118cb0e28af9dbd184bdb5f7e38))
+
+
+### Bug Fixes
+
+* lowercase kindfilter ([c6a8f67](https://github.com/socialgouv/kontinuous/commit/c6a8f67f4b12d8fba156607d59ed412cf5907189))
+* metabase version to 0.45.1 ([#119](https://github.com/socialgouv/kontinuous/issues/119)) ([148698d](https://github.com/socialgouv/kontinuous/commit/148698df4eabd094f02e42c857ae4210d6ca9ec9))
+* rename rolloutStatusManifest to more explicit ([e45df0f](https://github.com/socialgouv/kontinuous/commit/e45df0fdb7cda48a0c507ef0968addb2edf056ea))
+* upgrade rollout-status +add kind optimization ([529c91e](https://github.com/socialgouv/kontinuous/commit/529c91e0e010f725dc058a6a25cc78b2a145b11a))
+
+### 1.109.18 (2023-01-04)
+
+
+### Bug Fixes
+
+* metabase version to 0.45.1 ([#119](https://github.com/socialgouv/kontinuous/issues/119)) ([148698d](https://github.com/socialgouv/kontinuous/commit/148698df4eabd094f02e42c857ae4210d6ca9ec9))
+
+### [1.109.17](https://github.com/socialgouv/kontinuous/compare/v1.109.16...v1.109.17) (2023-01-04)
+
+
+### Bug Fixes
+
+* retry + logs ([19ea5fc](https://github.com/socialgouv/kontinuous/commit/19ea5fc9cdc056f71d278c0b97031e4363c65c6d))
+* various fixes ([211ead5](https://github.com/socialgouv/kontinuous/commit/211ead543d899c04903160014ba0c234e4a8f239))
+
+### [1.109.16](https://github.com/socialgouv/kontinuous/compare/v1.109.15...v1.109.16) (2023-01-04)
+
+
+### Bug Fixes
+
+* better handle of killed and error ([a42c8a8](https://github.com/socialgouv/kontinuous/commit/a42c8a80949d4228fc297ba9a1d74303778dd693))
+
+### 1.109.15 (2023-01-04)
+
+
+### Bug Fixes
+
+* **charts:** metabase upgrade to 0.45.1 and securityContext ([#116](https://github.com/socialgouv/kontinuous/issues/116)) ([facbbcd](https://github.com/socialgouv/kontinuous/commit/facbbcd2efc1ec4c7543da9796d79a94a5127cbd))
+
+### [1.109.14](https://github.com/socialgouv/kontinuous/compare/v1.109.13...v1.109.14) (2023-01-03)
+
+
+### Bug Fixes
+
+* queue pushAsync ([d58f7d0](https://github.com/socialgouv/kontinuous/commit/d58f7d038e088fb47b8ca1a751b8dbabd25736a2))
+
 ### [1.109.13](https://github.com/socialgouv/kontinuous/compare/v1.109.12...v1.109.13) (2022-12-18)